atualização questão 1

+++ src/atividade01/Questão06.java
@@ -6,10 +6,10 @@ public class Questão06 {
     public static void main(String[] args) {
         Scanner ler = new Scanner(System.in);
         System.out.println("Informar o valor da conta: ");
-        float valorConta = ler.nextFloat();
+        double valorConta = ler.nextDouble();
         System.out.println("Por quantas pessoas vai dividir a conta?: ");
         int divisaoConta = ler.nextInt();
-        float valorPessoa = (valorConta / divisaoConta);
+        double valorPessoa = (valorConta / divisaoConta);
         System.out.println("O valor que cada pessoa vai pagar é R$ "+valorPessoa+" reais");
         ler.close();
     }
